Ensure the memory for function names of entries are always valid
Getenv does not guarantee that the memory stays valid during the
further execution (getenv(3p)):

    The returned string pointer might be invalidated or the string
    content might be overwritten by a subsequent call to getenv(),
    setenv(), unsetenv(), or (if supported) putenv() but they shall
    not be affected by a call to any other function in this volume of
    POSIX.1‐2017.

So we create a copy, to ensure that the function names can be safely
accessed during later execution.  This may NOT be done using automatic
storage, since it is invalidated at the end of the current
context (function call).  Therefore we allocate static storage to
attempt and copy the configuration data and avoid the above mentioned
issue with getenv.