Ensure trip is always built as a position-independent executable
It seems that some systems (Debian) do this by default, which leads to
confusion, as others don't do this (Fedora).  By compiling with -fPIE
and binding with -pie, we ensure that the executable is usable as a
shared object as soon as "fix-pie" removes the PIE flag, that ld.so
checks in refusal to preload executables.