Check the GCC version before using -fhardened
Turns out that this would create a hard dependency on at least GCC 14,
which I don't want to do while Debian stable is still on GCC 12.  This
also means that we should check if the compiler is GCC at all.

Based on a trick[0] taken from the help-make mailing list.

[0] https://lists.gnu.org/archive/html/help-make/2006-04/msg00065.html