Always discard first PRNG value
When building trip without debugging options, the subsequent debugf
expression will not make use of n, which GCC complains about

trip.c:271:19: warning: unused variable ‘n’ [-Wunused-variable]
  271 |     unsigned long n = next();

Instead of checking if we are in debug mode, in which case we don't
trigger the PRNG, we add a noop usage of n.  This is both the
simplest solution, and avoids hypothetical issues between debug mode
and non-debug mode in terms of the number of PRNG invocations.