Fix sorting of know commands in 'names'
By using the default delimiters, blank spaces, the output or sort
wouldn't be correct, as e.g. "unlinkat," would appear before
"unlink,", due to the comma influencing the order.  The consequence of
this misordering was that bsearch could easily fail to find an element
in 'names', as the binary search would skip parts of the array,
assuming the element couldn't occur in there.  By delimiting at
commas, they do not affect the order and bsearch finds all elements as
intended.