Avoid unnecessary manual string termination for 'preload'
We know that readlink will terminate the string, if sufficient memory
was available.  As this condition is verified, we do not manually have
to ensure that the string is terminated any more.  This behaviour was
not the proper approach to begin with, seeing as an overflow would
have left us with an incorrect file path, that would have now become a
well-formed string (readlink will not append a NUL byte if the buffer
space was insufficient).

The assumption of the preceding loop is that readlink will have had
enough storage, hence will have had stored the proper result with a
NUL-byte.  So we assert this to be the case, hoping to uncover an
inconsistency if it arises, instead of executing a no-op by our
assumptions (overwriting what should be NUL with NUL).