Check if trip was installed successfully
As it is to be expected that people might try to install this script
without having made sure that ~/.local/bin is configure to be part of
their $PATH, it makes sense to print a constructive warning if that is
not the case, hinting what is to be done to resolve the issue.

The fallback of proposing to edit ~/.bashrc is certainly not ideal,
but has the same kind of user in mind.