Add -e option to list the errno values
It is currently a hassle to determine what possible `errno` values a
supported function can fail with, unless one opens the man page for
the respective function and search there. But even then, it is
ambiguous whether the `errno` values are from the The Open Group
Technical Standard Base Specifications (POSIX Standard), or from the
Linux man page.

To circumvent this, I have added an `-e` option that lists all the
`errno` values for a given function. If the function is not supported,
it prints an error message to `STDERR_FILENO` and exits with
`EXIT_FAILURE`. Else it prints each `errno` value on a separate line.

`strerrorname_np()` (a GNU C extension) is used to map an `errno`
value to its corresponding symbolic name. It was added in glibc 2.32,
which was released somewhere around 2020-08-05 according to
Sourceware.org. If code can not use it,then this StackOverflow answer
provides an alternative that can be used as a drop-in replacement:
[How can I print the symbolic name of an errno in
C?](https://stackoverflow.com/a/57519221/20017547).