Ensure that we never try to trip "trip" the executable
The linker can "apparently" choose to invoke our trampoline functions
while executing "trip" itself.  The consequence is that
'____trip_should_fail' and transitively 'chance' are invoked.  The
latter includes a "assert(is_lib)" check, which is not the case as we
set 'is_lib' to false when starting trip the executable.

The fix involves realising that we should not only not initialise the
library while trip the executable is running, but immediately return
from '____trip_should_fail' indicating that we don't want to trigger
an error.

This change relation to the issues described in
https://codeberg.org/pkal/trip/issues/1.