Do not pass a plain char to toupper().
According to the ISO C Standard:

> In all cases the argument in an `int`, the value of which
> shall be represented as an `unsigned char` or shall equal
> the value of the macro `EOF`. If the argument has any other
> value, the behavior is undefined.

The problem is that `char`, `signed char`, and `unsigned char`
are three distinct types, but `char` has the same range and
representation as *either* `signed char` or `unsigned char`.
(Plain `char` is very commonly signed and able to represent
values in the range -128..+127.)

If `char` is signed, then the call to `toupper()` would invoke
undefined behavior.

The fix is trivial, we only need a cast.