Replace GNU C's noreturn attribute with ISO C alternative.
ISO C11 has the `_Noreturn` keyword, along with a `noreturn`
convenience macro in `<stdnoreturn.h>`

A macro `noreturn` has been defined that works with all versions
of ISO C.

For C23 and above, it would expand to `[[noreturn]]` attribute
specifier sequence. For GCC, Clang, ICX, and other compiles
that claim compatibility with GNU C, it would expand to
`__attribute__(noreturn)`. For C11 and above, it would expand to
`_Noreturn`.