Fix use-after free in 'check_exec'
Since the we used to create a safe copy of the value of the
environmental variable PATH was allocated using strdup, and strtok
points (char *dir) into the storage allocated by it, we are not
allowed to free the memory before using it to prepare the command.

Instead we copy the value of PATH onto the stack and simplify the
current approach by prepending "." to the value, forcing the logic to
first check the current working directory, as was the case before with
the manual 'access' check.